backend/requirements.txt
# FastAPI and web framework dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sse-starlette>=1.8.0
python-multipart>=0.0.7

# Pydantic for data validation (included with FastAPI, but explicit version)
pydantic>=2.5.0

# Environment variables
python-dotenv>=1.0.0

# Inherited from original debate-ai requirements
# (these are needed because we import from src/)
anthropic>=0.18.0
rich>=13.0.0
